refactor(kvm): create 3-level pipeline skeleton and relocate v1 tasks
Implements the directory structure from pipeline plan v2 (#154):
- macro/ (6 UI action entry points)
- stages/ (9 stage orchestrators + 3 env contexts + common/)
- Relocated v1 tasks into stage-appropriate subdirectories
- Updated all imports to new paths
