+++ tools/pipeline/macro/demote.py
@@ -1,0 +1,1 @@
+"""Macro: demote a VM back to a lower zone."""
