+++ tools/pipeline/macro/destroy.py
@@ -1,0 +1,1 @@
+"""Macro: destroy a VM and clean up all resources."""
