+++ tools/pipeline/macro/promote_production.py
@@ -1,0 +1,1 @@
+"""Macro: promote a staging VM to production."""
