+++ tools/pipeline/macro/promote_staging.py
@@ -1,0 +1,1 @@
+"""Macro: promote a development VM to staging."""
