+++ tools/pipeline/macro/provision.py
@@ -1,0 +1,1 @@
+"""Macro: provision a new ERPNext VM (stages 1–9)."""
