+++ tools/pipeline/macro/refresh.py
@@ -1,0 +1,1 @@
+"""Macro: refresh an existing ERPNext VM (stages 3–9)."""
