+++ tools/pipeline/stages/common/types.py
