+++ tools/pipeline/stages/destroy.py
@@ -1,0 +1,1 @@
+"""Orchestrator: destroy a VM and clean up all resources."""
