+++ tools/pipeline/stages/env_cloudstack.py
@@ -1,0 +1,1 @@
+"""Environment context: CloudStack (iwStack VPS)."""
