+++ tools/pipeline/stages/env_vps.py
@@ -1,0 +1,1 @@
+"""Environment context: pre-existing VPS."""
