+++ tools/pipeline/stages/promote.py
@@ -1,0 +1,1 @@
+"""Orchestrator: promote a VM between zones."""
