+++ tools/pipeline/stages/stage_2_network/__init__.py
@@ -1,0 +1,1 @@
+"""Stage 2: Network — saconsole WG hub update, DNS, TLS cert distribution, WG spoke."""
