+++ tools/pipeline/stages/stage_3_connectivity/__init__.py
@@ -1,0 +1,1 @@
+"""Stage 3: Connectivity — deploy keys, ce_sri secrets, controller pubkey, backup."""
