+++ tools/pipeline/stages/stage_3_connectivity/controller_pubkey.py
@@ -4,8 +4,8 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from tools.pipeline.ssh import scp_to_vm, ssh_run
-from tools.pipeline.types import Config, Emit, TaskResult
+from tools.pipeline.stages.common.ssh import scp_to_vm, ssh_run
+from tools.pipeline.stages.common.types import Config, Emit, TaskResult
 
 CONTROLLER_PUBKEY = Path.home() / ".ssh" / "hasan_mighty.pub"
