+++ tools/pipeline/stages/stage_4_content_delivery/__init__.py
@@ -1,0 +1,1 @@
+"""Stage 4: Content Delivery — ddlViews, render config artifacts, deploy bundle."""
