+++ tools/pipeline/stages/stage_6_base_platform/__init__.py
@@ -1,0 +1,1 @@
+"""Stage 6: Base Platform — envars, bench rename, deploy keys, app clones, supervisor."""
