+++ tools/pipeline/stages/stage_7_data_restoration/__init__.py
@@ -1,0 +1,1 @@
+"""Stage 7: Data Restoration — new-site, install-app, handleRestore, fixture reimport."""
