+++ tools/pipeline/stages/stage_8_app_config/__init__.py
@@ -1,0 +1,1 @@
+"""Stage 8: App Config — install_specific before/after, secrets, .env generation."""
