+++ tools/pipeline/stages/stage_9_service_activation/__init__.py
@@ -1,0 +1,1 @@
+"""Stage 9: Service Activation — TLS deploy, Social Login, bash_aliases, final snapshot."""
